refactor: improve tutorial correctness and consistency
Use SDK's homograph normalizer instead of incomplete custom
implementation, centralize DPNS contract ID, rename legacy
dapiClientMethods to getSystemInfo, fix getDocument filename
to match plural function, validate network param in createClient,
and skip known-failing startAt test (dashpay/platform#3078).
