+++ test/queryTest.mjs
@@ -37,7 +37,8 @@ describe(`EVO SDK Query Tests (${new Date().toLocaleTimeString()})`, function su
   });
 
   describe('Query modifiers', function () {
-    it(`startAt - should return documents starting at document id - (${documentId})`, async function () {
+    // Skip: bare startAt query fails proof verification (dashpay/platform#3078)
+    it.skip(`startAt - should return documents starting at document id - (${documentId})`, async function () {
       const result = docsFromMap(await testQueries.startAt(sdk, documentId, 1));
 
       expect(result).to.have.lengthOf(1);
