+++ tutorials/constants.mjs
@@ -1,0 +1,3 @@
+// DPNS data contract ID for testnet
+// Find the current value at: https://docs.dash.org/projects/platform/en/stable/docs/reference/data-contracts.html
+export const DPNS_CONTRACT_ID = 'GWRSAVFMjXx8HpQFaNJMqBV7MBgMK4br5UESsB4S31Ec';
