+++ tutorials/index.mjs
@@ -1,9 +1,9 @@
 export { createClient } from './sdkClient.mjs';
 export { checkNetworkConnection } from './checkNetworkConnection.mjs';
-export { dapiClientMethods } from './dapiClientMethods.mjs';
+export { getSystemInfo } from './getSystemInfo.mjs';
 export { retrieveIdentity } from './identity/retrieveIdentity.mjs';
 export { retrieveNameByName } from './name/retrieveNameByName.mjs';
 export { retrieveNameByRecord } from './name/retrieveNameByRecord.mjs';
 export { retrieveNameBySearch } from './name/retrieveNameBySearch.mjs';
 export { retrieveContract } from './contract/retrieveContract.mjs';
-export { getDocuments } from './document/getDocument.mjs';
+export { getDocuments } from './document/getDocuments.mjs';
